# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Ted Charles Photography API
  description: |
    Backend API for Ted Charles Photography website providing:
    - Album and photo management
    - User authentication via Google OAuth
    - Site branding and configuration
    - Analytics event tracking
    - Public content serving (albums, photos, external links)
    
    ## Authentication
    Most endpoints are public. Admin endpoints require authentication via Google OAuth.
    The API uses session-based authentication with secure HTTP-only cookies.
    
    ## Security
    - CSRF protection on all mutating endpoints (POST/PUT/DELETE)
    - HMAC signature validation on analytics events
    - Path traversal protection on all file operations
    - Rate limiting on all endpoints
    
  version: 1.0.0
  contact:
    name: Ted Charles
    email: me@tedcharles.net
    url: https://tedcharles.net
  license:
    name: Creative Commons Attribution 4.0 International
    url: https://creativecommons.org/licenses/by/4.0/
servers:
  - url: https://api.tedcharles.net
    description: Production server
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /api/auth/google/callback:
    get:
      tags: [Authentication]
      summary: Google OAuth callback
      description: Handles the OAuth callback from Google
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: OAuth authorization code
      responses:
        '302':
          description: Redirect to admin portal on success, or error page on failure
...
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout
      description: Destroys the user session and logs out
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '401':
          description: Not authenticated
        '500':
          description: Logout or session destruction failed
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie set after Google OAuth authentication
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token for state-changing operations
...